Move most of the core functionality over to PDO
Since the mysql_* functions are being removed in PHP 5.6 and being
future-proof is generally a good idea this seemed like a good idea.

Any new functions added to the shared library should use PDO instead
of the old mysql functions and new features added to the projects
should preferably use the new pdoConnect() method to connect to the
database instead of dbConnect().