Fixed trigger
The trigger specified a table codeBox which doesn't exist in the database. Dropping the trigger completely causes a foreign key constraint failure when deleting a tuple from 'wordlist'. It seems apparent that 'codebox' is equivalent to the current 'box' table. Changing the table name from 'codeBox' to 'box' in the trigger seems to work and allows deletions from 'wordlist' without foreign key constraint errors.