Added email row in user
I've added a row in the user table that allows email addresses of users to be stored. For the time being, the attribute is declared to be default null so we don't run into issues with there already being users that have no email address, so either the database may need to be changed in the future or the constraint that a user must have a specified email needs to be controlled on the back end or front end.