Commented out incorrect foreign key constraints in studentresultat
The database would not build because the table studentresultat contained foreign key constraints that were not referencing primary or unique keys. I don't have design specifications for the database so I could not make the decision that certain rows should be set to unique. Therefore I commented out the faulty foreign key constraints.