Created a view to determine total hp for a user in a specific course
I've removed the previous trigger solution because it seemed like it would turn real ugly fast. Making it easier to find out how many credits a student has in a course by using a view is a really clean solution. This commit includes comments above the new view that demonstrate how the view could be used.