Added styling to <h1> when in descboxes
As the code is now, h1 (when in a box) gets white, bold, italic and centered. That is partly done in ".box h1"-class. To make description text more uniform I addded styling to h1 when in descboxes. Changing the styling back to the "original" look defined in "style.css".