Fixed browser resizing problems.
The problem boiled down to jQuery having a prefrence of doing everything pixelbased, and the previous functions were set to work on exact measurments in pixels. However this is now resolved, all measurments as far as i can see have been converted to percentage based ones. Which solves the browser resizing problem.