Fixed issue #1938
Changed the code so it would be able to handle stuff between numbered
lists without breaking the couting sequence. Also changed how lists are
implemented. This also fixes sublists for an ordered list. Added to
JavaScript and PHP.
(###) To start an ordered list.
1. Hi (number, dot, space, and content to start couting sequence).
(/###) To end an ordered list.
To create sublists just start a list within another list.