Bugs, animation and clarity of collapsable sections
Fixed the bug where the sections are created into div's without classes.
Separated content for "header" and "section". Also added an animation
when the section collaps. "Sections" are now collapsable to. "Moments"
and "Sections" now have a pointer as a cursor and arrows to clearify
that they are clickable to collapse the content related to them.