Changed FAB and New Item button in sectioned.php
Before the FAB created a set Section. The user had to click on the cog wheel to change it everytime the FAB was used. Now the Edit Item View appears when FAB is clicked which means that a set Section is not
created everytime. The user has to choose which type of dugga should be created directly after clicking the FAB.