Made the comment work again after weekly merge
The comment has been fixed after weekly merge. Now the comment will be saved in the database again and will be shown on the website.  Also added a control of what kind of item is created and chosen. The comment option will only show if the item that is chosen is a Test.