Re-commit
Rownumber already exists and works as a magic header. Styling such as
borders etc won't be implemented as of now, would like to know more
specifically how they want it styled or if the solution is even remotely
close to what they want before I begin styling (as this could take quite
some time given the way the elements are created).