Added all groups in grouped.php
All groups that exists are now shown on grouped.php. It is better to show groups instead of all students. The students will be shown beside the groupname which makes it easy to see which student is placed in the groups. Show the students together with the groups will be a later issue.