Added logic to display a notification(challenge)
Function has been added to set localstorage "securitynotification" which
is used to display or show the notification when the document is loaded.
Some HTML markup has been added for the notification window. Security
question is now also fetched from the database and a session is set
based on if it's empty or not. Localstorage is set depending on if the
session is null or not.