Fixed to the group editor lists per student instead of per group
Made it easier to connect a student to a list entry and added stoplight
to better reflect if a student is assigned to a list entry. Added a
unique identifier to all content cells to be able to identify the cell
with userid, course id and listentry id. Those are needed to be able to
toggle group access in the list later, which there is also a
placeholder function for now.