Made some more sorting changes in grouped.php
It is now possible to order students by firstname, lastname, username and ssn in grouped.php. This is done by hover on the icon to the right in the navheader. Some styling has also been deleted from style.css that was not supposed to have been added. It was added by mistake when issue #3681 was merged in to this issue.