Fix-For-Issue-#3321-&-#3400-&-#3409
 - Figures can now be selected on their own. Either by draging and selecting all points of the figure or by clicking inside the figure.
 - Figures can now be moved on their own.
 - Figures can now be reformed/resized. (a single point of a figure is now movable)
 - Figures can now be deleted.


NOTE: Inconsistency when selecting figures. Some times it only works if you drag and select all points of the figure, and sometimes it only works if you click inside of the figure. Eg. to remove a figure you have to drag and select all points.

Open a new issue about this!