Issue #3733 & #3766
#3766: Cuts out the dialog window functionality into a seperate javascript file.

#3733:
Created a function for loading the forms with a simple AJAX-script.

This also gives the editing within the diagram more visual feedback. It laso removes a lot of redundant code by making functions more streamline.

ID:s and inner variables AttributeType, EntityType and RelationType have been removed and replaced with "object_type".

Fixed a bug where double-clicking anywhere in the diagram (that isn't an object) causes an error.