Fix for issue #3691
This solves the problem with the OK-button.

Someone should take a look at the styling/media queries for the button
for desktop view, that one does not look great, it hangs a bit too much
to the right. But for mobile devices it looks good now, and the issue
with the button is solved.