Fix for issue #3802
I've now fixed so that when you doubleclick on a symbol, it will load the current data from the array into the forms. I've also added some padding to the Name-box. 

NOTE: When you doubleclick, the console will say that there is an error. But there is no error. This is a bug for a later issue.