Functionallity has been added to the funnel button in grouped.php
If a teacher wants to see all moments in the dropdownlist in the funnel button then all moments will be shown in the table in grouped.php. If one moment is checked, only that moment will be shown in the table for example. This makes the funnel icon work as intended.