Makes 'name' in usergroup not UNIQUE
Changed 'name' in 'usergroup' table to be not unique anymore. The reason is because usergroups are connected to lid(listentry id)  and therefore it should be possible to have the same name on groups, but the groups still need to have different lid to be able to have the same name. If 'name' would be unique it would not be possible to have 2 groups with the name 'A' for example.