Makes the sorting work if filtering has been done in grouped.php
There was a bug where the sorting stops working if filtering has been done, the sorting only broke if the moments was shown in a different order than it was default. For example, if Biträkningsdugga is not shown the table will still be sorting on that moment because it was shown as the first moment as default. This commit solves that bug which makes the sorting work if some moments are filtered and not shown.