Removed the static height of input element of 26px
For some reason there is a set height for ALL inputs in the LenaSYS. Removed this hardcoded input height enforcement, and if someone needs to set a height for a specific input, they should state this separatly for that specific input.
There were not set height at start of the project.