Dropdownlist will be hidden if mouse hover Back-button in grouped.php
The dropdownlist that appears when mouse hover over Funnel icon will now disappear when the mouse hover over Back-button in navheader. This does mean that the dropdownlist will not be shown incorrectly anymore. The same issue was also present in resulted.php but it will also be solved with this change.