#4180_ver01
Changed the dugga date size from static width to an adaptive width by adding a new css class called ”dateSize”. This class is later used in different media queries in order to be able to change the width of the date size based on the screen size.

Changed the dugga title in order to be able to wrap the text based on screen size and free space. The task was competed by adding ”nowrap” and ”ellipsis”  classes on the titles div tag.

Changed the font size for the course title to be able to adapt based on screen size. This was completed by changing the font size in the different media queries that the ”style” CSS file used.