Local storage for menus work
The state of the collapsed menus are saved in local storage and remembered when reloading the page.

Possible bug:
Sometimes the state of the menu is remembered but not visible on the page, i.e. collapsed menus appear open. It is probably related to when the function for hidding elements is called.