#4208_ver01
Fixed the issue when writing headings, sections, etc. that include quotes (") in sectioned and duggaed will make editing of that line impossible.

The task was completed by creating a method that checks if the title input has any invalid characters. If any invalid characters are found then the input row changes color to red, and the submit/save button cant be pressed until all invalid characters are gone.