#5027_ver1
Generalized the previous "No Test"-dialog window, so the FAB chioces "Code" and "Link" also can use it to display it's error message, by just changing the error message text through an id with jQuery.

Added if-statments for kind 2 and 5 to validate if code examples or links exists in the database for that specific course. Code Examples has a hard codes index, which makes the statment different for the others.

Finally, added code to close the dialog window by pressing "Enter".