#5029_ver1
Removed the hard coded tooltip message in sectioned.php, so it can be added via jQuery instead - depending on what kind is validated.

Added else-statments to the isTypeValid() function so it can check if any links or code examples exist under Files. If not, a message will be displayd using jQuery for that specific kind/type.

Additional, I removed some breaklines and added spaces where it was needed where I was adding my functionality.