#5159_Fix
Changed the buttons so that the "Send" button says "create" to make it clear that it creates a new variant. The "save" button also has a changed name to "update" to clarify that it updates a variation.

To duplicate a variation it is necessary to click edit on a variant and then press the "create" button. Making multiple duplicates at once does not work well because the table will be updated between each individual insertion.

This solution is no doubt enough workable when the maximum number of variants for a dugga is 36.