Tabel sorting fix for the access, test and files tab.
This is a solution that solves the sorting problem for the tables in access, test and  files tab. Instead of using 3 different compare functions i made one compare function that all sortingTables uses. The compare function is universal and can be used for more tables later on.

This solution works for the tabels at the pages stated above but do not work for the results table under the "Results tab" on the course front page.

My suggestion is that we take this solution and then adding the results table as an new issue and trying to solve that one with this solution.