Moved comment blok in accessed.js
Undid the comment covering 2 functions to instead comment away the content of the functions. This avoids the uncaught reference error whenever the functions got called (were never removed and can't really be removed) and it avoids displaying an empty white box. This should fix issue #6037 and #6052 whilst keeping the fix on issue #5975