Update diagram_objects.js
I have fixed so the cardinality dosen´t gets hidden by the weak entity. But when the entity is in a diagonal with the relation object the position of the cardinality gets closer to the line. But if you move the entity object down the cardinality will get to it´s right position