Now changes cursor when hovering
hoveredObject and markedObject replaces hovobj as hovobj was set to the marked object.

Lines are added after objects in diagram so they will be hovered instead of an attribute even though it is drawn behind it. This has to do with the lines being added to diagram afterwards but being connected to the center of attributes rather than to the edge.