Added line "uimode = "CreateLine"  to diagram.js
Drawing a UML-line changes uimode to "CreateUMLLine", but if the line is not drawn uimode is never changed back to "ChangeLine" and thus does not pass any checks until the button is pressed again and uimode is changed back. This was solved by simply adding the line "uimode = "CreateLine";" where the check occurs if the line is drawn to an empty spot. This enables you to draw a new line to another class.