Alignment with free draw objects fix
According to issue #5920 alignment with free drawn objects is broken. comments mentions that an exeption for path is probably the best solution here. This exeption was implemented, aligning now should move all object accordingly while ignoring free draw objects.

Snap to grid was also broken on free draw objects, so a fix was made so that free draw object now also snaps to grid when moving it instead of breaking like before.