Fixed the reference error in the console
Fixed it so there will no longer be an error in the console because "fabBtn" is not defined (or rather, the js.file loads in before the button loads in and therefor it can't find it, but now it can). This also fixes the problem with the Deadlines" tab not closing and opening. It works now.