Fixed the fab buttons on the files page in mobile view
Fixed the fab buttons on the files page in mobile view so that they show up when you press the +. You also need to only press the +, and not anywhere on the screen.

You can still close the fab buttons by clicking on the background but that is not an issue as it's usually how things are on mobile.