Icon redesign to a more common and modern version
The icon has been changed to a more common icon for filtering.

Another icon is included but not used, this icon is the same as the original icon but the ring has been removed to optimize it to look more modern. It's there in case it's more desired than the current new choice.

The new icon used is filter_icon.svg
The optimised unused icon is tratt_v2.svg