On screen message handling
A function called "flash" has been implemented which shows a danger/succes message on screen.

This function is now used to show a flash message to the user if he/she tries to make a free draw object with less than 3 points(makes two points then right-clicks).

This function can further be used to show any type of message which may be useful to the user.