#7690 Solved issue where written feedback gets deleted by session expiration.
Solved the issue by storing the text in the feedback textarea to localstorage when the session expires. And later retrieve the text when opening the feedback textarea again.

Also solved the problem with the page not reloading when getting logged out.

I wasn't able to properly test these changes. So be extra hard on the review to be sure that it all works as it's supposed to.