Exporting as SVG now handles UMl lines
Solves #7543
When exporting diagram to SVG the UML lines should now be displayed, and in hopefully all cases be displayed exactly how they appear in the diagram.

Edit: due to how SVG export currently works, the captured area will always be from from camera-position 0,0 even though you've moved the camera. This messed up the use of breakpoints and was solved by simply moving the camera to position 0,0 before svg export. Maybe not the best solution, but i belive a new issue should address this if needed.