Download all now works
Possible to download as intented for all files from a specific version as mentioned in issue#7765. It does create a zip on the server that is not removed and that could be a possible issue later on. There has to be a verifaction from the server to the client that the file has been downloaded before it can be removed and that is why we left it for a later issue. Also made it so only teachers or people with write access and super users can downlod the files because that was also a critera of the issue#7765.

Co-Authored-By: b18emigu <b18emigu@users.noreply.github.com>