Solution for copying ER-lines when copy-pasting
Fixes #6977

A solution which makes it possible for ER-lines to be copied and pasted, while keeping correct connections to it's copied objects.

Note: UML lines ar more diffecult to deal with due to breakpoints, and currently uml classes can't even be copied. So this should be made to a sepparate issue.