#8287 Fixed issue were duggas was unasigned still when a sub-dugga has been handed in.
Added a for loop to connect duggor with its sub-duggor.

Made a variable to save the placeholder grade. 0 if  a sub-dugga has been handed in and Null if no sub-dugga has been handed in.

Made changes to when how many times a dugga has been marked and time when submitted.

Co-Authored-By: a18matna <a18matna@users.noreply.github.com>