Made changes to parameters and lots of other stuff
The reason that the 'tabs' option did not do anything was because there was no space in the database where that value was stored.

I added a new parameter in sectionedservice.php which connected to the local changes I made in my local database via phpmyadmin, where I added a field in 'listentries' for tabs.

Now that there is a field for tabs, the value can be stored and called upon when the website is loaded or changed so that the indentation of the tabs can change according to what option we choose.

Added the class "tabs + item["tabs"]" which gives each <td> with the right id of "indTab" the value it's been changed to in the "tabs" select option. The class will be changed to what 'tab' was selected, and depending on what id it has (0,1,2,3,4, etc), different styling is applied to make indentations.

For this to work on the real webpage there needs to be an additional field in the database for 'tabs', as I have added to my local database.