#1934 Added so you can submit userfeedback on duggor.
Added function to send an AJAX call with the values from the userfeedback form.

The AJAX call insert them in the database and based on if the user wants to be contacted the user id also is stored otherwise not to keep anonymity.

Also added status messeges when submiting form.

Co-Authored-By: a18matna <a18matna@users.noreply.github.com>