Fix for issue #8741
Fixes #8741
Fixed issue by changing selection that decides when UML-line should be drawn or not. Also moved when uimode is changed to outside the selection.

This fixes the bug with the odd line being drawn when incorrect line draw is attempted(ER to UML). This also fixes a other issue when drawLine would break and need to be toggled on and off for it to work again when incorrect drawLine is attempted. This is probably due to uimode before being toggled inside selection and would be unreachable.

Test by drawing a mix of correct and incorrect lines while also creating new objects.