Solves issue #8765
Before a dummy point was created when drawing a line from a freedraw-object to a ER-object. This was due to freedraw-objects not being handled in the "okToMakeLine" function.

This also fixed a error only visible in the console when drawing a line to a freedraw-object.

Test by drawing lines from and to free-drawobjects. No errors should occur and no dummy-points should be created.