Removed unused functions
The functions are not currently used, if the need for something similar arrives the functions would have to look very different anyway. This is due to new additions and the fact that the old functions do not follow code standard. All currently active dropdowns should still work. All manus should work as well.