Solves #8553  - Solution for connecting loose lines to objects
Dragging a loose line end into a object will now connect them if connection between objects is valid.

Line should not connect if connection is invalid(example: two lines between same ER-attributes).

Note for tester:
Loose line is easiest created by copy-pasting a ER-line. You can then drag one of the lines endpoints inside a object, and it should connect if connection is valid.

Also note that it only works for ER-lines for now. This is due to copying and pasting UML-lines not working correctly as of now, and therefore these lines are not handled yet.