#1934 Moved GUI for checking userfeedback to sectioned. Also fixed AJAX call to get the userfeedback.
Moved the GUI for checking userfeedback to sectioned because it will make it easier for the teacher to see it.

Fixed so when you open the GUI a AJAX call is made to get the userfeedback from the database.

Started working on a table to show the userfeedback.

Changed the UID attribute in the userfeedback table to Username.

Co-Authored-By: a18matna <a18matna@users.noreply.github.com>