#8859 Create an 'new announcement' button is only accessed by the admin
Default button is removed.
The button is created and displayed when the admin logs in.
Created an input element in navheader because passing php variables value to javascript does not seem to work here, so I had to find an another solution.
Now if you login as 'brom'  the button is display and hidden when logged in as 'teacher2'