Small fixes
Fixed a small error when it comes to the if-statement that was made for #8905.

There have been a lot of investigating to see where the date string gets lost (it appears to be UNK in the end). I has been narrowed down to a problem with AJAX. This will be further investigated.

Co-Authored-By: a18eripi <a18eripi@users.noreply.github.com>