Fix for #9066
Special case has been created for when both objects connected to a line is attributes. When this is the case we do not create new points.

Test by connecting two attributes with a line, then copying and pasting them. May also be worth connecting other combinations of objects to make sure nothing else broke.