More changes made
Had to merge this branch with the W20 branch one more time because there were some peroblems.

The "616px" for the FABStatic2 button are back to "auto" now., as it should be. Also added a "left:auto" for MOTDbutton so that it doesn't float to the right of the screen.