Working solution for manually changing  the side of symbol a line is connected to
It is now possible to manually change which side of a symbol which a line connects to. Do this by opening the appearance menu for a line, then use the two dropdowns under "line placement" to change the side of the corresponding symbol.

Known issues which will be created sepperatly if pull-request goes through:
 - Changing sides of recursive UML lines does not behave correctly.
-Loose lines are not handled. Not a huge issue, but may be implemented if customer needs it.