Time to live now fixed for variants
Set the default time to live to 30 days (2592000000 in milliseconds). Every time one adds a localstorage variable and they wish to add a ttl to it, they need to use the function setExpireTime which is located within dugga.js (row circa line 450). Each time a localstorage is retrieved, use getExpireTime to check on the time to live and if the item should be removed from localstorage.