Changed the position divider
Changed the position divider to a lower value and removed the constant to make it scale better for longer lists.

Should work fine for 70+ items.

There is still a "bugg" where this function sometimes struggle when there are alot of newly added items. It fixes itself if you leave the site and rejoins later.

Co-Authored-By: b19wilfo <62884368+b19wilfo@users.noreply.github.com>