Updated the placement of new items
Updated the code that makes sure that new items are placed at appropriate places and not off-screen. We still have a bug that places new items at the button at the screen off too many items are created at the same time. We are still waiting for response from teacher regarding this bug.

Co-author: @a19danla