Restored the functionality of the back and forward buttons respecting the order the examples are placed in.
Moved the part where you check if a template has been applied to the example to the top so that:
1: a user entering an example that does not have a template applied to it gets an alert that states the error rather than the webpage crashing.

2: A teacher teacher that enters an example gets prompted to add a template instead of the webpage crashing.

Previously all the examples with a type['kind'] of 2 were placed in a seperate array called codeExamples,
This was prone to bugs so now the code pulls what position they are currently in the list of examples from sectionData(defined at the top of the returned function).

sectionData includes examples that are of kind 1 and so some checks are made to avoid redirecting the user to a page that is not a code example when filling in the before and after arrays.