Updated fab in style.css
Changed transition in fab-btn-list when window is <= 600px in height. The text now doesn't overlap and is obliquely downwards instead of upwards. Also moved the fab-button so the text will always be vissible.

Test this by logging in as mestr, navigate to "show files" and hover over the fab button (the yellow large round one). Then make the window smaller and hover over it again.