Created UML entity with placeholder for attributes
Added attributes to the UML element (Attributes can not be changed yet but the box containing the attributes expands). You can also now click on the entity and the normal properties appear. An potential issue is that some code has been moved in the drawElements() function. This is due to support the special creation of the UML entity. I have tried to minimize refactor to the bare minimum.