Updated renderCommits in contribution.js
Updated renderCommits to take into account that the data may not have the same year as the current year(2022). Instead made it so it checks if the commit happened the same year as the first date in data. Also added some text to show the user the interval date for the commit tree.