Fixed fab-btn not staying up when hovering
Fixed the issue with not being able to keep fab-btn up when hovering. The issue appears to be a a miscalculation. the fab-btn got a margin of 30 but the fab-btn-list got a margin increase of 35 which meant there probably was a gap of 5 pixels which would count as no longer hovering over the element thus hiding the option.