Added dark mode (toggle) on two more files.
Added dark mode (toggle) on two more files. In the "codewiver.php" file, the code had a different type of structure than that implemented in sectioned.php. There is a comment that explains it.

I also put back "media prefers-color-scheme" on both css files. Leave a comment that they should stay there and not be removed.