Made the sidebar only show for superusers
The sidebar for accepting users is no longer created in the contribution.php file. Instead the sidebar is made in contribution.js. Now when a guest visits contribution.php the sidebar will not be shown. After a user has loged in if they are a superuser the sidebar will be created.