Options panel updates dynamically
The content of the option panel changes depending on the selected element. Yuo can aslo change between diagram types on elements that has a ER and UML variant. The content in the other element is still saved. Functions and attributes are added as you type in the textareas. New line, new entry.

TODO: Filter out empty entries, allow none entries in either attribute or functions. Remove the need to click the element again when changing the diagram type.