Fixed onclick function in contribution.php
Fixed the onclick function for the rankTable and the ghContribTable. The previous version used the same rendersSortOptions as myTable which meant that the function were trying to sort on the wrong table. Also had to add the compare function which is called to sort the table since it was missing from contribution.js. The compare works for sorting ints and strings but not for arrays so i added the functionality of handling those arrays to the compare function.