Fixed broken show selected dates in contribution
The wrong userid was used for the ajax which meant that instead of getting the current user the ajax was sending the option selected in <select id="userid> button which is  "select Git user". Also fixed a problem with the "showing activities for period" string displaying undefined dates if the user hadn't selected any dates for a user.