added more commitDiv functionality
Changed commitDiv from being hidden or shown based on mouseover to instead onclick.
Moved the github link to a newly added icon instead so that the onclick functionality could be on the text. Made clicking on anything outside of the commitDiv hide the element.
Made the icons change to display: none; when the contribContent is closed and display: inline-flex; when open.