Template 9 reszising now functional.
It should now be possible to resize the boxes in template 9.
I will add that the increased (human)complexity is due to there being 4 boxes stacked on top of each other in this template.
If there is a better way to do this i implore the person reading to do that.