Fixed courseed tab order
Fixed the taborder of the elements on courseed. Moved fab btn code so it will be after the course list and not before it. Also made it possible to perform the onclick functionality of the fab-btn using only keyboard. Changed the css style so that all elements in the course list are clearly marker when tabbed onto.