found how to separate the titles from the rest of the drop down
By using selector `span[id$="examples"]` all items in the array which has an id ending with "examples" are selected, which happen to be all section titles that groups the code examples in one course.  The selector can be use to style the elements separately from the other items in list.