Added mobile support for copy button
Added a 0.5 second delay to the selectioncahnge event because when mobile users have selected code and press the cody button their selection is deselected before the copy command is executed. With the slight delay that will not be a problem. Also added clipboard  API support to replace depricated method execCommand().