Rework of current functionality
Since the chaos from Friday, some part of the code has been reworked with comments and easier code. The main change is that the weak keys are now created as they should however what is left is to work on how the weak keys are passed as foreign in different relation and also to print out the values in the ER table.