Added save and load functionality to SVG dugga
Added the save button functionality to store the content from the editors into the userAnswer table and also for that data to be retrieved later by a superUser.
Also changed how showDuggaservice loads data from userAnswer. previously it always used moment as a key but now if the hash matches it will use that instead. This should work better with load dugga functionality. Had to add unset($grade) because it prevented submissions since grade was "UNK" by default which isset and greater than 1 making submissions impossible.