Recursively calling a github repo
What we've done is fetch the api-link for a specific github repo to retrieve the data in JSON-format with fetch (await fetch(url)), afterwards we run a breadth first traversal to go through the data including sub-folders of the original data. The out-commented code below console.log(files) is a previously used algorithm to further dive into the data and structure them into individual arrays, but it's in need of refactoring to fit into the current fully functioning version of the fetch with analysis.

Done w/ @e21krida

Co-Authored-By: e21krida <102578849+e21krida@users.noreply.github.com>