Update diagram.php
found an inconsistency in getVariantParam where some of the php variables that were pushed into the array were surrounded by '' while others weren't. This caused the code to be parsed in a way that made the browser think UNK wasn't just a placeholder string but an actual variable name that was, of course, undefined.