Fixed comments and error-handling
Implemented a simple error-handling through checking if $data is true or false, since this is the response that is sent if the fetch is successful by file_get_contents. The default error-message is removed by adding a @ in front of the "file_get_contents" as this avoids default error messages. Also added error-handling in case the $json returns as incorrect/false.

Removed two files, "recursivetesting\recursivejs.js" & "recursivetesting\recursivephp.php", these are both supposed to be removed from the final weekly branch as these were temporary testing files from previous renditions of this solution.