Changed alerts to console.log().
Due to Async calls the console logs was printed in the console before the Async call was completed for all the objects. Therefore we had to wrap the whole Async call into a Promise and then use the Promise from setInterval to ensure that the Async call was completed in its entirety before logging the success.