Update navheader.php
It was trying to create a button in the hamburger menu with parameters from a version of a course that is not yet created. There was a if-statement to check if versname from the response from table vers is existing for the button in the navbar but not the hamburger menu. Placed the button in the hamburger menu in a new if-statement that checks if versname is existing aswell.
Also removed dead code that was used for test an error from a previous issue.

Co-Authored-By: a18chrkr <102579106+a18chrkr@users.noreply.github.com>