Changed the implementation of some colorMap calls
Removed the const colorKeys variabel. Changed the calls of colorMap to aquire the keys and values with the Object class instead. See changed lines: 

7185-7197
7238-7239

Altough a function creation was necessary since Objects doesnt have a set index like an array, and the index of the color had to be fetched aswell as the colorcode, see line 7197.