Updated insertIntoSQLite() function
Updated insertIntoSQLite() function by adding code which splits the received URL into different parts, and extracting the 4th section  which contains the name of the repo. It is then assigned to the variable repoName. Bindparam is used so that repoName can be used in the now modified SQL query and will consequently insert the repoName (which was not part of the query previously).